MAINT: Move `LazySequenceLearner` to separate module
